refactor(s3): rename s3.py to backend.py for uniform module layout (M1)
Every other backend's AbstractDataSource subclass lives in backend.py; move
S3's there too. Only the package __init__ imported earthlens.s3.s3 (public
access is via earthlens.s3), so no shim is needed.
