+++ src/earthlens/s3/__init__.py
@@ -12,6 +12,6 @@ from __future__ import annotations
 
 from earthlens.s3.auth import S3Auth, S3Credentials
 from earthlens.s3.catalog import Catalog, Dataset, Variable
-from earthlens.s3.s3 import S3
+from earthlens.s3.backend import S3
 
 __all__ = ["S3", "Catalog", "Dataset", "Variable", "S3Auth", "S3Credentials"]
